# openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "FastAPI",
                 "version":  "0.1.0"
             },
    "paths":  {
...
                  "/api/interview/questions":  {
                                                   "get":  {
                                                               "tags":  [
                                                                            "interview"
                                                                        ],
                                                               "summary":  "Get Interview Questions",
                                                               "operationId":  "get_interview_questions_api_interview_questions_get",
                                                               "parameters":  [
                                                                                  {
                                                                                      "name":  "company",
                                                                                      "in":  "query",
                                                                                      "required":  false,
                                                                                      "schema":  {
                                                                                                     "anyOf":  " ",
                                                                                                     "title":  "Company"
                                                                                                 }
                                                                                  },
                                                                                  {
                                                                                      "name":  "role",
                                                                                      "in":  "query",
                                                                                      "required":  false,
                                                                                      "schema":  {
                                                                                                     "anyOf":  " ",
                                                                                                     "title":  "Role"
                                                                                                 }
                                                                                  },
                                                                                  {
                                                                                      "name":  "experience",
                                                                                      "in":  "query",
                                                                                      "required":  false,
                                                                                      "schema":  {
                                                                                                     "anyOf":  " ",
                                                                                                     "title":  "Experience"
                                                                                                 }
                                                                                  },
                                                                                  {
                                                                                      "name":  "session",
                                                                                      "in":  "query",
                                                                                      "required":  false,
                                                                                      "schema":  {
                                                                                                     "anyOf":  " ",
                                                                                                     "title":  "Session"
                                                                                                 }
                                                                                  },
                                                                                  {
                                                                                      "name":  "no_repeat_days",
                                                                                      "in":  "query",
                                                                                      "required":  false,
                                                                                      "schema":  {
                                                                                                     "type":  "integer",
                                                                                                     "maximum":  365,
                                                                                                     "minimum":  1,
                                                                                                     "default":  90,
                                                                                                     "title":  "No Repeat Days"
                                                                                                 }
                                                                                  }
                                                                              ],
                                                               "responses":  {
                                                                                 "200":  {
                                                                                             "description":  "Successful Response",
                                                                                             "content":  {
                                                                                                             "application/json":  "@{schema=}"
                                                                                                         }
                                                                                         },
                                                                                 "422":  {
                                                                                             "description":  "Validation Error",
                                                                                             "content":  {
                                                                                                             "application/json":  "@{schema=}"
                                                                                                         }
                                                                                         }
                                                                             }
                                                           }
                                               },
                  "/api/interview/start-with-jd":  {
                                                       "post":  {
                                                                    "tags":  [
                                                                                 "interview"
                                                                             ],
                                                                    "summary":  "Start Interview With Jd",
                                                                    "description":  "1. Analyze a Job Description (JD) using local AI service.\n2. Extract Company and Role (Experience Level).\n3. Fetch one matching question from the question pool.",
                                                                    "operationId":  "start_interview_with_jd_api_interview_start_with_jd_post",
                                                                    "requestBody":  {
                                                                                        "content":  {
                                                                                                        "application/json":  {
                                                                                                                                 "schema":  "@{$ref=#/components/schemas/JDUpload}"
                                                                                                                             }
                                                                                                    },
                                                                                        "required":  true
                                                                                    },
                                                                    "responses":  {
                                                                                      "200":  {
                                                                                                  "description":  "Successful Response",
                                                                                                  "content":  {
                                                                                                                  "application/json":  "@{schema=}"
                                                                                                              }
                                                                                              },
                                                                                      "422":  {
                                                                                                  "description":  "Validation Error",
                                                                                                  "content":  {
                                                                                                                  "application/json":  "@{schema=}"
                                                                                                              }
                                                                                              }
                                                                                  }
                                                                }
                                                   },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "JDUpload":  {
                                                        "properties":  {
                                                                           "jd_text":  {
                                                                                           "type":  "string",
                                                                                           "title":  "Jd Text"
                                                                                       }
                                                                       },
                                                        "type":  "object",
                                                        "required":  [
                                                                         "jd_text"
                                                                     ],
                                                        "title":  "JDUpload",
                                                        "description":  "Input schema for starting an interview with a JD upload."
                                                    },
...
                                   }
                   }
}